Move components under right pages
