+++ src/app/users/[id]/profileInfo.tsx
@@ -5,7 +5,7 @@ import { getServerSession } from "next-auth";
 import { authOptions } from "@/authOptions";
 import { getDescriptionAndSocialMedia } from "@/services/userService";
 import { ProfileButtons } from "./profileButtons";
-import { Friend, FriendListButton } from "@/components/friendListButton";
+import { type Friend, FriendListButton } from "./friendListButton";
 
 export const ProfileInfo = async ({ userId, friends }: { userId: number; friends: Friend[] }) => {
   const session = await getServerSession(authOptions);
